requirements.txt
alembic==1.7.7
amqp==5.1.1
anyio==3.6.1
apache-airflow==2.3.1
apache-airflow-providers-celery==2.1.4
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-google==7.0.0
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==2.0.0
attrs==20.3.0
awscli==1.24.9
Babel==2.10.1
billiard==3.6.4.0
blinker==1.4
boto3==1.23.8
botocore==1.26.9
cachelib==0.7.0
cachetools==5.1.0
cattrs==1.10.0
celery==5.2.6
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.1
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==36.0.2
db-dtypes==1.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
Flask==1.1.2
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flower==1.0.0
future==0.18.2
google-ads==16.0.0
google-api-core==2.8.0
google-api-python-client==1.12.11
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud-aiplatform==1.13.0
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-automl==2.7.2
google-cloud-bigquery==2.34.3
google-cloud-bigquery-datatransfer==3.6.1
google-cloud-bigquery-storage==2.13.1
google-cloud-bigtable==1.7.1
google-cloud-build==3.8.2
google-cloud-container==2.10.7
google-cloud-core==2.3.0
google-cloud-datacatalog==3.8.0
google-cloud-dataplex==1.0.0
google-cloud-dataproc==4.0.2
google-cloud-dataproc-metastore==1.5.0
google-cloud-dlp==1.0.1
google-cloud-kms==2.11.1
google-cloud-language==1.3.1
google-cloud-logging==3.1.1
google-cloud-memcache==1.3.1
google-cloud-monitoring==2.9.1
google-cloud-orchestration-airflow==1.3.1
google-cloud-os-login==2.6.1
google-cloud-pubsub==2.12.1
google-cloud-redis==2.8.0
google-cloud-resource-manager==1.5.0
google-cloud-secret-manager==1.0.1
google-cloud-spanner==1.19.2
google-cloud-speech==1.3.3
google-cloud-storage==1.44.0
google-cloud-tasks==2.9.0
google-cloud-texttospeech==1.0.2
google-cloud-translate==1.7.1
google-cloud-videointelligence==1.16.2
google-cloud-vision==1.0.1
google-cloud-workflows==1.6.2
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.1
graphviz==0.20
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.46.3
grpcio-gcp==0.2.2
grpcio-status==1.46.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
humanize==4.1.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==3.0.3
jmespath==1.0.0
json-merge-patch==0.2
jsonschema==4.5.1
kombu==5.2.4
lazy-object-proxy==1.7.1
lockfile==0.12.2
looker-sdk==22.4.0
lxml==4.8.0
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
multitasking==0.0.10
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandas-gbq==0.17.5
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.19.6
protobuf==3.20.0
psutil==5.9.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.4.0
Pygments==2.12.0
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.4.1
rsa==4.7.2
s3transfer==0.5.2
sentry-sdk==1.5.12
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.9
sqlalchemy-bigquery==1.4.3
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
statsd==3.3.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.1
typing_extensions==4.2.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
yfinance==0.1.70
zipp==3.8.0
